Fix email update not working.
Fix an issue where we failed to include our own should-process logic.  The update-gsuser was then using it's own, which is backwards to what ours is.  IE, it always confirms instead of only confirming when passing the -Confirm flag.  We usually bypass this by doing our own should-process logic.  However, this section of code was not done and was not working when someone's email was changed.